Refactor scoring related tests
